Weapon Working
무기 슬롯에 스크립트를 넣어놨으니, 추후 다시 확인하여 작업을 유지할 것

+++ ActionMake00/Assets/Scripts/Buff/BuffBase.cs
@@ -53,7 +53,7 @@ public abstract class BuffBase
     /// <param name="target"></param>
     /// <param name="dmgAmount"></param>
     /// <param name="duration"></param>
-    public virtual void ObjectSetup(Character target, int dmgAmount, float duration)
+    public virtual void ObjectSetup(Character target, float dmgAmount, float duration)
     {
         character = target;
         Init(duration, ApplyBuff, UpdateBuff, ExitBuff);
